There's a chance this will fix CI.
