+++ ci/cibuildwheel.yaml
@@ -69,9 +69,6 @@ stages:
               set -o errexit
               python3 -m pip install --upgrade pip
               python3 -m pip install cibuildwheel==2.11.2
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
             displayName: Install dependencies
           - bash: cibuildwheel --output-dir wheelhouse .
             displayName: Build wheels
@@ -87,9 +84,6 @@ stages:
               set -o errexit
               python3 -m pip install --upgrade pip
               python3 -m pip install cibuildwheel==2.11.2
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
             displayName: Install dependencies
           - bash: cibuildwheel --output-dir wheelhouse .
             displayName: Build wheels
@@ -107,9 +101,6 @@ stages:
               set -o errexit
               python3 -m pip install --upgrade pip
               python3 -m pip install cibuildwheel==2.11.2
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
             displayName: Install dependencies
           - bash: cibuildwheel --output-dir wheelhouse .
             displayName: Build wheels
@@ -127,9 +118,6 @@ stages:
               set -o errexit
               python3 -m pip install --upgrade pip
               python3 -m pip install cibuildwheel==2.11.2
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
             displayName: Install dependencies
           - bash: cibuildwheel --output-dir wheelhouse .
             displayName: Build wheels
@@ -147,9 +135,6 @@ stages:
               set -o errexit
               python3 -m pip install --upgrade pip
               python3 -m pip install cibuildwheel==2.11.2
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
             displayName: Install dependencies
           - bash: cibuildwheel --output-dir wheelhouse .
             displayName: Build wheels
@@ -167,9 +152,6 @@ stages:
               set -o errexit
               python3 -m pip install --upgrade pip
               python3 -m pip install cibuildwheel==2.11.2
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
             displayName: Install dependencies
           - bash: cibuildwheel --output-dir wheelhouse .
             displayName: Build wheels
@@ -185,9 +167,6 @@ stages:
               set -o errexit
               python3 -m pip install --upgrade pip
               python3 -m pip install cibuildwheel==2.11.2
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
             displayName: Install dependencies
           - bash: cibuildwheel --output-dir wheelhouse .
             displayName: Build wheels
