+++ ci/run_tests_pipeline.yaml
@@ -30,9 +30,7 @@ stages:
           - script: python3 --version
           - script: |
               python3 -m pip install cython
-              python3 -m pip install numpy
-              python3 -m pip install pandas
-              python3 -m pip install pyarrow
+              python3 ci/pip_install_deps.py
             displayName: Installing Python dependencies
           - script: python3 proj.py build
             displayName: "Build"
