+++ pyproject.toml
@@ -51,10 +51,7 @@ requires = [
 # See: https://cibuildwheel.readthedocs.io/en/stable/options/#configuration-file
 build-verbosity = "3"
 before-build = "python {project}/install_rust.py"
-before-test = """
-    python -m pip install numpy>=1.21.6
-    python -m pip install pandas>=1.3.5
-    """
+before-test = "python {project}/ci/pip_install_deps.py"
 test-command = "python {project}/test/test.py -v"
 skip = [
     # No 32-bit musl C native tool chain for Rust.
